+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # ChangeLog
 
 
+## 1.4.0 (August 27, 2019)
+- Enhancement: Added width/height props to `<Video />` [#7](https://github.com/ihmpavel/expo-video-player/issues/7)
+- Enhancement: Added `videoBackground` prop
+- Enhancement: Added more video examples
+- Enhancement: Checking internet status with hooks
+
 ## 1.3.1 (August 25, 2019)
 - Fix: Do not add to npm registry Lint source files
 
@@ -8,8 +14,8 @@
 - Enhancement: Added changelog
 - Enhancement: Fully rewritten, code cleanup
 - Enhancement: Rewritten `example app`, now in TS
-- Fix: Add setAudioModeAsync key ([#2](https://github.com/ihmpavel/expo-video-player/issues/2)
-- Fix: Updated Expo ([#3](https://github.com/ihmpavel/expo-video-player/issues/3)
+- Fix: Add setAudioModeAsync key [#2](https://github.com/ihmpavel/expo-video-player/issues/2)
+- Fix: Updated Expo [#3](https://github.com/ihmpavel/expo-video-player/issues/3)
 
 ## 1.2.0 (April 12, 2019)
 - Enhancement: Refactored code